Create TestQ2.php
Everyday, the restaurant has a limited number of breads, vadas and samosas. The
restaurant needs to optimise orders in a way that they can generate maximum profit.
Vadapav needs two bread and a vada
Samosapav needs two bread and a samosa
Input:
First input is: no. of breads, no. of vadas and no. of samosas available in the
restaurant
Second input is: Price of Vadapav & price of Samosapav set by the restaurant
Output:
Maximum profit possible
Example:
Input:
9 2 3 (9 breads, 2 vada, 3 samosa)
10 15 (Rs.10 - Vadapav, Rs.15 - Samosapav)
Output:
55 (Maximum profit possible is Rs.55)